--- src/test/java/com/tobe/healthy/schedule/StandByScheduleTest.java
@@ -1,25 +1,0 @@
-package com.tobe.healthy.schedule;
-
-import com.tobe.healthy.schedule.domain.entity.Schedule;
-import jakarta.persistence.EntityManager;
-import lombok.extern.slf4j.Slf4j;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.transaction.annotation.Transactional;
-
-@Slf4j
-@SpringBootTest
-@Transactional
-class StandByScheduleTest {
-
-	@Autowired
-	private EntityManager em;
-
-	@Test
-	void 예약_대기를_한다() {
-		Schedule schedule = em.find(Schedule.class, 1L);
-		log.info("schedule => {}", schedule);
-
-	}
-}
